requirements.txt
django-crispy-forms==1.14.0
Pillow==10.0.1
pandas==2.1.3
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0